+++ Sources/Sentry/SentryHub.m
@@ -18,6 +18,7 @@
 #import "SentryScope+Private.h"
 #import "SentrySerialization.h"
 #import "SentrySession+Private.h"
+#import "SentryTraceOrigins.h"
 #import "SentryTracer.h"
 #import "SentryTracesSampler.h"
 #import "SentryTransaction.h"
@@ -110,7 +111,6 @@ SentryHub ()
         [scope applyToSession:_session];
 
         [self storeCurrentSession:_session];
-        // TODO: Capture outside the lock. Not the reference in the scope.
         [self captureSession:_session];
     }
     [lastSession endSessionExitedWithTimestamp:[self.currentDateProvider date]];
@@ -306,7 +306,8 @@ SentryHub ()
     return [self startTransactionWithContext:[[SentryTransactionContext alloc]
                                                  initWithName:name
                                                    nameSource:kSentryTransactionNameSourceCustom
-                                                    operation:operation]];
+                                                    operation:operation
+                                                       origin:SentryTraceOriginManual]];
 }
 
 - (id<SentrySpan>)startTransactionWithName:(NSString *)name
@@ -316,7 +317,8 @@ SentryHub ()
     return [self startTransactionWithContext:[[SentryTransactionContext alloc]
                                                  initWithName:name
                                                    nameSource:kSentryTransactionNameSourceCustom
-                                                    operation:operation]
+                                                    operation:operation
+                                                       origin:SentryTraceOriginManual]
                                  bindToScope:bindToScope];
 }
 
@@ -358,6 +360,7 @@ SentryHub ()
     return [[SentryTransactionContext alloc] initWithName:context.name
                                                nameSource:context.nameSource
                                                 operation:context.operation
+                                                   origin:context.origin
                                                   traceId:context.traceId
                                                    spanId:context.spanId
                                              parentSpanId:context.parentSpanId