+++ Sources/Sentry/SentryStacktraceBuilder.m
@@ -45,7 +45,6 @@ SentryStacktraceBuilder ()
             [frames addObject:frame];
         }
     }
-    sentrycrash_async_backtrace_decref(stackCursor.async_caller);
 
     NSArray<SentryFrame *> *framesCleared = [SentryFrameRemover removeNonSdkFrames:frames];