+++ Sources/Sentry/include/SentryProfileTimeseries.h
@@ -8,13 +8,6 @@
 
 @class SentryTransaction;
 
-/**
- * Synchronizes reads and writes to the samples array; otherwise there will be a data race between
- * when the sampling profiler tries to insert a new sample, and when we iterate over the sample
- * array with fast enumeration to extract only those samples needed for a given transaction.
- */
-SENTRY_EXTERN std::mutex _gSamplesArrayLock;
-
 NS_ASSUME_NONNULL_BEGIN
 
 /** A storage class to hold the data associated with a single profiler sample. */