+++ Tests/SentryTests/Integrations/Performance/CoreData/SentryCoreDataTrackerTest.swift
@@ -64,7 +64,7 @@ class SentryCoreDataTrackerTests: XCTestCase {
             expect.fulfill()
         }
 
-        wait(for: [expect], timeout: 0.1)
+        wait(for: [expect], timeout: 1.0)
     }
     
     func test_FetchRequest_WithPredicate() {
@@ -111,7 +111,7 @@ class SentryCoreDataTrackerTests: XCTestCase {
             expect.fulfill()
         }
 
-        wait(for: [expect], timeout: 0.1)
+        wait(for: [expect], timeout: 1.0)
     }
     
     func test_Save_2Insert_1Entity() {
@@ -307,6 +307,7 @@ class SentryCoreDataTrackerTests: XCTestCase {
 
         XCTAssertEqual(result?.count, 1)
         XCTAssertEqual(dbSpan.operation, SENTRY_COREDATA_FETCH_OPERATION)
+        XCTAssertEqual(dbSpan.origin, "auto.db.core_data")
         XCTAssertEqual(dbSpan.spanDescription, expectedDescription)
         XCTAssertEqual(dbSpan.data["read_count"] as? Int, 1)
         XCTAssertEqual(dbSpan.data["blocked_main_thread"] as? Bool ?? false, mainThread)